+++ tests/commands/test_modify.py
@@ -12,7 +12,7 @@ import pytest
 from typing_extensions import override
 
 from cobib.commands import ModifyCommand
-from cobib.config import Event
+from cobib.config import Event, config
 from cobib.database import Database
 from cobib.utils.rel_path import RelPath
 
@@ -170,13 +170,19 @@ class TestModifyCommand(CommandTest):
             assert Database()[expected].label == expected
 
     @pytest.mark.parametrize("preserve_files", [True, False])
-    def test_rename_associated_file(self, setup: Any, preserve_files: bool) -> None:
+    @pytest.mark.parametrize("config_overwrite", [True, False])
+    def test_rename_associated_file(
+        self, setup: Any, preserve_files: bool, config_overwrite: bool
+    ) -> None:
         """Test removing associated files.
 
         Args:
             setup: the `tests.commands.command_test.CommandTest.setup` fixture.
-            preserve_files: argument to `DeleteCommand`.
+            preserve_files: argument to `ModifyCommand`.
+            config_overwrite: what to overwrite `config.commands.modify.preserve_files` with.
         """
+        config.commands.modify.preserve_files = config_overwrite
+
         with tempfile.TemporaryDirectory() as tmpdirname:
             path = RelPath(tmpdirname + "/knuthwebsite.pdf")
             open(path.path, "w", encoding="utf-8").close()  # pylint: disable=consider-using-with
@@ -190,7 +196,7 @@ class TestModifyCommand(CommandTest):
             assert "dummy" in Database().keys()
 
             target = RelPath(tmpdirname + "/dummy.pdf")
-            if preserve_files:
+            if preserve_files or config_overwrite:
                 assert path.path.exists()
             else:
                 assert target.path.exists()